backend/requirements.txt
sentence-transformers==3.1.1
pandas==2.2.3
numpy==2.1.1
langchain_openai==0.2.1
networkx==3.3
python-dotenv==1.0.1